+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to Semantic Versioning (https://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- `pcb bom` again colors planner-selected parts by green, yellow, and red sourceability tiers.
+
 ## [0.4.31] - 2026-08-18
 
 ### Added